Bugfix: filter out DNI components within a group
So far the script assumed all componets in a group(=BOM line) would have the same
DNI status. Which is true if the 'value' field is set to DNI, but nog if
the DNI/DNP property is used!

For example, if one item in a group (=same footprint and value) is marked DNI
but the rest is not, it would still be included on the BOM. Or the whole
BOM line would be removed if it was the last part in the group!